+++ ngClient/gulpfile.js
@@ -7,11 +7,13 @@ var gulp = require('gulp'),
     cssmin = require('gulp-cssmin'),
     jshint = require('gulp-jshint'),
     rename = require('gulp-rename'),
+    sourcemaps = require('gulp-sourcemaps'),
     tap = require('gulp-tap'),
     uglify = require('gulp-uglify');
 
 // app.js variables
 var appJsRoot = './js/app',
+    appJsSourceMapRoot = appJsRoot.substring(1),
     appJs = 'app.js',
     appJsPath = appJsRoot + '/' + appJs,
     appMinJs = 'app.min.js',
@@ -27,38 +29,40 @@ var appCssRoot = './css',
     appCssSrc = [appCssRoot + '/*.css', appJsRoot + '/directives/**/*.css', '!' + appCssPath, '!' + appMinCssPath];
 
 // lib variables
-var libSrcRoot = './bower_components',
+var libJsSrcRoot = './bower_components',
+    libJsSourceMapRoot = libJsSrcRoot.substring(1),
 
     // Js
-    jquery = libSrcRoot + '/jquery/dist/jquery.js',
-    respond = libSrcRoot + '/respond/dest/respond.js',
-    datajs = libSrcRoot + '/datajs/datajs.js',
-    toastr = libSrcRoot + '/toastr/toastr.js',
-    angular = libSrcRoot + '/angular/angular.js',
-    angularRoute = libSrcRoot + '/angular-route/angular-route.js',
-    angularSanitize = libSrcRoot + '/angular-sanitize/angular-sanitize.js',
-    breeze = libSrcRoot + '/breeze-client/build/breeze.base.debug.js',
-    breezeAjaxAngular = libSrcRoot + '/breeze-client/build/adapters/breeze.ajax.angular.js',
-    breezeBridgeAngular = libSrcRoot + '/breeze-client/build/adapters/breeze.bridge.angular.js',
+    jquery = libJsSrcRoot + '/jquery/dist/jquery.js',
+    angular = libJsSrcRoot + '/angular/angular.js',
+    angularRoute = libJsSrcRoot + '/angular-route/angular-route.js',
+    angularSanitize = libJsSrcRoot + '/angular-sanitize/angular-sanitize.js',
+    datajs = libJsSrcRoot + '/datajs/datajs.js',
+    breeze = libJsSrcRoot + '/breeze-client/build/breeze.base.debug.js',
+    breezeAjaxAngular = libJsSrcRoot + '/breeze-client/build/adapters/breeze.ajax.angular.js',
     breezeDataServiceOData = './js/lib/breeze-client/build/adapters/breeze.dataService.odata.js', // Fixed version
-    breezeModelLibraryBackingStore = libSrcRoot + '/breeze-client/build/adapters/breeze.modelLibrary.backingStore.js',
-    breezeUriBuilderOData = libSrcRoot + '/breeze-client/build/adapters/breeze.uriBuilder.odata.js',
-    breezeDirectives = libSrcRoot + '/breeze-client-labs/breeze.directives.js',
-    googleAnalyticsAngular = libSrcRoot + '/angular-google-analytics/dist/angular-google-analytics.js',
-    disqusAngular = libSrcRoot + '/angular-utils-disqus/dirDisqus.js',
-    bootstrapAngular = libSrcRoot + '/angular-bootstrap/ui-bootstrap-tpls.js',
-    highcharts = libSrcRoot + '/highcharts/highcharts.src.js',
-    highchartsAngular = libSrcRoot + '/highcharts-ng/dist/highcharts-ng.js',
+    breezeModelLibraryBackingStore = libJsSrcRoot + '/breeze-client/build/adapters/breeze.modelLibrary.backingStore.js',
+    breezeUriBuilderOData = libJsSrcRoot + '/breeze-client/build/adapters/breeze.uriBuilder.odata.js',
+    breezeBridgeAngular = libJsSrcRoot + '/breeze-client/build/adapters/breeze.bridge.angular.js',
+    breezeDirectives = libJsSrcRoot + '/breeze-client-labs/breeze.directives.js',
+    googleAnalyticsAngular = libJsSrcRoot + '/angular-google-analytics/dist/angular-google-analytics.js',
+    disqusAngular = libJsSrcRoot + '/angular-utils-disqus/dirDisqus.js',
+    respond = libJsSrcRoot + '/respond/dest/respond.js',
+    bootstrapAngular = libJsSrcRoot + '/angular-bootstrap/ui-bootstrap-tpls.js',
+    highcharts = libJsSrcRoot + '/highcharts/highcharts.src.js',
+    highchartsAngular = libJsSrcRoot + '/highcharts-ng/dist/highcharts-ng.js',
+    toastr = libJsSrcRoot + '/toastr/toastr.js',
+    sourceMap = libJsSrcRoot + '/source-map/dist/source-map.js',
 
     // Css
-    bootstrap = libSrcRoot + '/bootstrap/dist/css/bootstrap.css',
-    fontAwesome = libSrcRoot + '/font-awesome/css/font-awesome.css',
-    breezeDirectivesCss = libSrcRoot + '/breeze-client-labs/breeze.directives.css',
-    toastrCss = libSrcRoot + '/toastr/toastr.css',
-    bootstrapSocial = '/bootstrap-social/bootstrap-social.css',
+    bootstrap = libJsSrcRoot + '/bootstrap/dist/css/bootstrap.css',
+    fontAwesome = libJsSrcRoot + '/font-awesome/css/font-awesome.css',
+    bootstrapSocial = libJsSrcRoot + '/bootstrap-social/bootstrap-social.css',
+    breezeDirectivesCss = libJsSrcRoot + '/breeze-client-labs/breeze.directives.css',
+    toastrCss = libJsSrcRoot + '/toastr/toastr.css',
 
     // Fonts (Font Awesome)
-    fontAwesomeFonts = libSrcRoot + '/font-awesome/fonts/*',
+    fontAwesomeFonts = libJsSrcRoot + '/font-awesome/fonts/*',
     fontAwesomeFontsDest = './css/fonts';
 
 // lib.js variables
@@ -67,7 +71,9 @@ var libJsRoot = './js/lib',
     libJsPath = libJsRoot + '/' + libJs,
     libMinJs = 'lib.min.js',
     libMinJsPath = libJsRoot + '/' + libMinJs,
-    libJsSrc = [jquery, respond, datajs, toastr, angular, angularRoute, angularSanitize, breeze, breezeAjaxAngular, breezeDataServiceOData, breezeModelLibraryBackingStore, breezeUriBuilderOData, breezeBridgeAngular, breezeDirectives, googleAnalyticsAngular, disqusAngular, bootstrapAngular, highcharts, highchartsAngular];
+    libJsSrc = [jquery, angular, angularRoute, angularSanitize, datajs, breeze, breezeAjaxAngular, breezeDataServiceOData,
+        breezeModelLibraryBackingStore, breezeUriBuilderOData, breezeBridgeAngular, breezeDirectives, googleAnalyticsAngular,
+        disqusAngular, respond, bootstrapAngular, highcharts, highchartsAngular, toastr, sourceMap];
 
 // lib.css variables
 var libCssRoot = './css/lib',
@@ -75,7 +81,7 @@ var libCssRoot = './css/lib',
     libCssPath = libCssRoot + '/' + libCss,
     libMinCss = 'lib.min.css',
     libMinCssPath = libCssRoot + '/' + libMinCss,
-    libCssSrc = [bootstrap, fontAwesome, breezeDirectivesCss, toastrCss, bootstrapSocial];
+    libCssSrc = [bootstrap, fontAwesome, bootstrapSocial, breezeDirectivesCss, toastrCss];
 
 // default
 gulp.task('default', [appJs, appCss, libJs, libCss, 'watch']);
@@ -86,10 +92,13 @@ gulp.task(appJs, function () {
     return gulp.src(appJsSrc)
         .pipe(jshint())
         .pipe(jshint.reporter('jshint-stylish'))
+        .pipe(sourcemaps.init())
         .pipe(concat(appJs, { newLine: '\r\n' }))
         .pipe(gulp.dest(appJsRoot))
         .pipe(rename(appMinJs))
         .pipe(uglify())
+        .on('error', errorHandler)
+        .pipe(sourcemaps.write('./', { sourceRoot: appJsSourceMapRoot }))
         .pipe(gulp.dest(appJsRoot));
 });
 
@@ -108,11 +117,15 @@ gulp.task(appCss, function () {
 gulp.task(libJs, function () {
 
     return gulp.src(libJsSrc)
+        .pipe(sourcemaps.init())
         .pipe(concat(libJs, { newLine: '\r\n' }))
         .pipe(gulp.dest(libJsRoot))
         .pipe(rename(libMinJs))
         .pipe(uglify())
-        .pipe(gulp.dest(libJsRoot));
+        .on('error', errorHandler)
+        .pipe(sourcemaps.write('./', { sourceRoot: libJsSourceMapRoot }))
+        .pipe(gulp.dest(libJsRoot))
+    ;
 });
 
 // lib.css: copy font awesome fonts + concat all into lib.css + minify all into lib.min.css
@@ -137,3 +150,8 @@ gulp.task('watch', function () {
     gulp.watch(libCssSrc, [libCss]);
 
 });
+
+function errorHandler(error) {
+    console.log(error);
+    this.emit('end');
+}